Refactors the imports
