+++ src/reducers/authSlice.ts
@@ -13,4 +13,4 @@ export const authSlice = createSlice({
 });
 
 export const { setIsUserLoggedIn } = authSlice.actions;
-export default authSlice.reducer;
+export const authReducer = authSlice.reducer;
