+++ src/reducers/index.ts
@@ -1,0 +1,2 @@
+export * from './authSlice';
+export * from './appointmentSlice';
